IoP vex: fix tests
